feat: report download progress
there is no way to track download sequentially
as s5cmd run or cp command will be locked on a thread.
so, download progress is outsourced to a process
and will run simultaneously along side download
process for both download from selection and manifest

For manifest or selection, download size
is calculated first and the progress is tracked
against as a whole.

manifest validator in download from manifest is now
offloaded to a dedicated function that can check
not only the first line but every line

when sync dry run is enabled, s5cmd cp is replaced with sync to
gracefully avoid downloading the same data again.

add additional endpoints download_dicom_studies, download_dicom_patients,
download_dicom_patients and download_collection, all of them routed to
download_from_selection

Co-Authored-By: Andrey Fedorov <andrey.fedorov@gmail.com>
Co-Authored-By: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>